+++ custom_components/zaptec/config_flow.py
@@ -1,14 +1,13 @@
 """Adds config flow for zaptec."""
 from __future__ import annotations
 
-import http
 import logging
 
 import aiohttp
 import voluptuous as vol
 from homeassistant import config_entries
 
-from .api import Account, AuthorizationFailedException
+from .api import Account, AuthorizationError
 from .const import DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
@@ -54,7 +53,7 @@ class ZaptecFlowHandler(config_entries.ConfigFlow, domain=DOMAIN):
                 )
             except aiohttp.ClientConnectorError:
                 errors["base"] = "connection_failure"
-            except AuthorizationFailedException:
+            except AuthorizationError:
                 errors["base"] = "auth_failure"
 
             if valid_login: