+++ custom_components/zaptec/services.py
@@ -1,19 +1,21 @@
 """Zaptec components services."""
 from __future__ import annotations
 
-import asyncio
 import logging
-import os
+from typing import Awaitable, Callable
 
 import voluptuous as vol
 from homeassistant.core import HomeAssistant, ServiceCall
 
-from .api import Account
+from .api import Account, Charger, Installation
 from .const import DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
 
+TServiceHandler = Callable[[ServiceCall], Awaitable[None]]
 
+# SCHEMAS for services
+# ====================
 has_id_schema = vol.Schema({vol.Required("charger_id"): str})
 
 has_limit_current_schema = vol.Schema(vol.SomeOf(
@@ -32,102 +34,75 @@ has_limit_current_schema = vol.Schema(vol.SomeOf(
     },
 ]))
 
-has_redacted_schema = vol.Schema({vol.Optional("redacted", default=True): bool})
 
-
-async def NO_async_setup_services(hass: HomeAssistant) -> None:
+async def async_setup_services(hass: HomeAssistant) -> None:
     """Set up services for the Plex component."""
 
     _LOGGER.debug("Set up services")
     acc: Account = hass.data[DOMAIN]["api"]
 
-    # just the new one for now.
-    # #  require firmware > 3.2
     async def service_handle_stop_pause(service_call: ServiceCall) -> None:
-        _LOGGER.debug("called new stop pause")
+        _LOGGER.debug("Called stop pause")
         charger_id = service_call.data["charger_id"]
-        return await acc.map[charger_id].command("stop_pause")
+        charger: Charger = acc.map[charger_id]
+        await charger.command("stop_pause")
 
     async def service_handle_resume_charging(service_call: ServiceCall) -> None:
-        _LOGGER.debug("service new start and or resume")
+        _LOGGER.debug("Called start and or resume")
         charger_id = service_call.data["charger_id"]
-        return await acc.map[charger_id].command("resume_charging")
+        charger: Charger = acc.map[charger_id]
+        await charger.command("resume_charging")
 
-    # Add old one to see if they even work.
-    async def service_handle_start_charging(service_call: ServiceCall) -> None:
-        _LOGGER.debug("service old start")
+    async def service_handle_authorize_charging(service_call: ServiceCall) -> None:
+        _LOGGER.debug("Called authorize charging")
         charger_id = service_call.data["charger_id"]
-        cmd = f"chargers/{charger_id}/SendCommand/501"
-        return await acc._request(cmd, method="post")
+        charger: Charger = acc.map[charger_id]
+        await charger.command("authorize_charge")
 
-    async def service_handle_stop_charging(service_call: ServiceCall) -> None:
-        _LOGGER.debug("service old stop")
+    async def service_handle_deauthorize_charging(service_call: ServiceCall) -> None:
+        _LOGGER.debug("Called deauthorize charging and stop")
         charger_id = service_call.data["charger_id"]
-        cmd = f"chargers/{charger_id}/SendCommand/502"
-        return await acc._request(cmd, method="post")
+        charger: Charger = acc.map[charger_id]
+        await charger.command("deauthorize_stop")
 
     async def service_handle_restart_charger(service_call: ServiceCall) -> None:
-        _LOGGER.debug("service restart_charger")
+        _LOGGER.debug("Called restart charger")
         charger_id = service_call.data["charger_id"]
-        return await acc.map[charger_id].command("restart_charger")
+        charger: Charger = acc.map[charger_id]
+        await charger.command("restart_charger")
 
     async def service_handle_update_firmware(service_call: ServiceCall) -> None:
-        _LOGGER.debug("service update_firmware")
+        _LOGGER.debug("Called update firmware")
         charger_id = service_call.data["charger_id"]
-        return await acc.map[charger_id].update_firmware()
+        charger: Charger = acc.map[charger_id]
+        await charger.command("upgrade_firmware")
 
     async def service_handle_limit_current(service_call: ServiceCall) -> None:
-        _LOGGER.debug("update current limit")
+        _LOGGER.debug("Called set current limit")
         installation_id = service_call.data["installation_id"]
         available_current = service_call.data.get("available_current")
         available_current_phase1 = service_call.data.get("available_current_phase1")
         available_current_phase2 = service_call.data.get("available_current_phase2")
         available_current_phase3 = service_call.data.get("available_current_phase3")
-        return await acc.map[installation_id].limit_current(
+        installation: Installation = acc.map[installation_id]
+        await installation.limit_current(
             availableCurrent=available_current,
             availableCurrentPhase1=available_current_phase1,
             availableCurrentPhase2=available_current_phase2,
             availableCurrentPhase3=available_current_phase3,
         )
 
-    async def service_handle_debug_data_dump(service_call: ServiceCall) -> None:
-        _LOGGER.debug("debug data dump")
-        redacted = service_call.data["redacted"]
-        path = os.path.join(hass.config.config_dir, 'www', 'zaptec')
-        os.makedirs(path, exist_ok=True)
-        with open(os.path.join(path, 'debug.txt'), 'w') as f:
-            async for text in acc.data_dump(redacted=redacted):
-                f.write(text)
-        _LOGGER.warning("Dumped Zaptec debug info. Restart HA and download from <URL>/local/zaptec/debug.txt")
+    # LIST OF SERVICES
+    services: list[tuple[str, vol.Schema, TServiceHandler]] = [
+        ("stop_pause_charging",  has_id_schema, service_handle_stop_pause),
+        ("resume_charging",      has_id_schema, service_handle_resume_charging),
+        ("authorize_charging",   has_id_schema, service_handle_authorize_charging),
+        ("deauthorize_charging", has_id_schema, service_handle_deauthorize_charging),
+        ("restart_charger",      has_id_schema, service_handle_restart_charger),
+        ("update_firmware",      has_id_schema, service_handle_update_firmware),
+        ("limit_current",        has_limit_current_schema, service_handle_limit_current),
+    ]
 
-    hass.services.async_register(
-        DOMAIN, "stop_pause_charging", service_handle_stop_pause, schema=has_id_schema
-    )
-
-    hass.services.async_register(
-        DOMAIN, "resume_charging", service_handle_resume_charging, schema=has_id_schema
-    )
-
-    hass.services.async_register(
-        DOMAIN, "start_charging", service_handle_start_charging, schema=has_id_schema
-    )
-
-    hass.services.async_register(
-        DOMAIN, "stop_charging", service_handle_stop_charging, schema=has_id_schema
-    )
-
-    hass.services.async_register(
-        DOMAIN, "restart_charger", service_handle_restart_charger, schema=has_id_schema
-    )
-
-    hass.services.async_register(
-        DOMAIN, "update_firmware", service_handle_update_firmware, schema=has_id_schema
-    )
-
-    hass.services.async_register(
-        DOMAIN, "limit_current", service_handle_limit_current, schema=has_limit_current_schema
-    )
-
-    hass.services.async_register(
-        DOMAIN, "debug_data_dump", service_handle_debug_data_dump, schema=has_redacted_schema
-    )
+    # Register the services
+    for name, schema, handler in services:
+        hass.services.async_register(DOMAIN, name, handler, schema=schema)